backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
faiss-cpu>=1.12.0
pydantic==2.5.3
python-multipart==0.0.6
anthropic>=0.18.0
python-dotenv==1.0.0
